Fix compiler warning.

+++ Library/Source/Nano/Threads/NThreadPool.cpp
@@ -432,7 +432,7 @@ void NThreadPool::ScheduleTask(NThreadTask *theTask)
 
 	// Validate our state
 	NN_ASSERT(mLock.IsLocked());
-	NN_ASSERT(mTasks.size() < kMaxTasks);
+	NN_ASSERT(mTasks.size() < (size_t) kMaxTasks);
